requirements.txt
ansible-core==2.14
Jinja2==3.1.2
pynetbox==6.6.2
pytz==2022.6
ansible-pylibssh==1.0.0
textfsm==1.1.3
python-dotenv==0.15.0
